refactor: DTO 이너클래스화
