--- be/src/main/java/kr/codesquad/secondhand/api/member/dto/request/MemberProfileImgUpdateRequest.java
@@ -1,14 +1,0 @@
-package kr.codesquad.secondhand.api.member.dto.request;
-
-import lombok.Getter;
-import org.springframework.web.multipart.MultipartFile;
-
-@Getter
-public class MemberProfileImgUpdateRequest {
-
-    private MultipartFile newProfileImage;
-
-    public MemberProfileImgUpdateRequest(MultipartFile newProfileImage) {
-        this.newProfileImage = newProfileImage;
-    }
-}
