+++ be/src/main/java/kr/codesquad/secondhand/api/member/dto/request/MemberRequest.java
@@ -1,0 +1,17 @@
+package kr.codesquad.secondhand.api.member.dto.request;
+
+import lombok.Getter;
+import org.springframework.web.multipart.MultipartFile;
+
+public class MemberRequest {
+
+    @Getter
+    public static class MemberProfileImgUpdateRequest {
+
+        private MultipartFile newProfileImage;
+
+        public MemberProfileImgUpdateRequest(MultipartFile newProfileImage) {
+            this.newProfileImage = newProfileImage;
+        }
+    }
+}
