--- be/src/main/java/kr/codesquad/secondhand/api/member/dto/response/MemberProfileImgUrlResponse.java
@@ -1,13 +1,0 @@
-package kr.codesquad.secondhand.api.member.dto.response;
-
-import lombok.Getter;
-
-@Getter
-public class MemberProfileImgUrlResponse {
-
-    private String updatedImgUrl;
-
-    public MemberProfileImgUrlResponse(String updatedImgUrl) {
-        this.updatedImgUrl = updatedImgUrl;
-    }
-}
