+++ be/src/main/java/kr/codesquad/secondhand/api/member/dto/response/MemberResponse.java
@@ -1,0 +1,17 @@
+package kr.codesquad.secondhand.api.member.dto.response;
+
+import lombok.Getter;
+
+public class MemberResponse {
+
+    @Getter
+    public static class MemberProfileImgUpdateResponse {
+
+        private String updatedImgUrl;
+
+        public MemberProfileImgUpdateResponse(String updatedImgUrl) {
+            this.updatedImgUrl = updatedImgUrl;
+        }
+    }
+
+}
