+++ be/src/main/java/kr/codesquad/secondhand/api/member/service/MemberFacadeService.java
@@ -6,9 +6,9 @@ import kr.codesquad.secondhand.api.address.service.AddressService;
 import kr.codesquad.secondhand.api.jwt.domain.Jwt;
 import kr.codesquad.secondhand.api.jwt.service.JwtService;
 import kr.codesquad.secondhand.api.member.domain.Member;
-import kr.codesquad.secondhand.api.member.dto.request.MemberProfileImgUpdateRequest;
+import kr.codesquad.secondhand.api.member.dto.request.MemberRequest.MemberProfileImgUpdateRequest;
 import kr.codesquad.secondhand.api.member.dto.response.MemberAddressResponse;
-import kr.codesquad.secondhand.api.member.dto.response.MemberProfileImgUrlResponse;
+import kr.codesquad.secondhand.api.member.dto.response.MemberResponse.MemberProfileImgUpdateResponse;
 import kr.codesquad.secondhand.api.member.dto.response.OAuthSignInResponse;
 import kr.codesquad.secondhand.api.oauth.domain.OAuthProfile;
 import kr.codesquad.secondhand.api.oauth.service.OAuthService;
@@ -44,11 +44,11 @@ public class MemberFacadeService {
     }
 
     @Transactional
-    public MemberProfileImgUrlResponse updateMemberProfileImg(Long memberId,
-                                                              MemberProfileImgUpdateRequest memberProfileImgUpdateRequest) {
+    public MemberProfileImgUpdateResponse updateMemberProfileImg(Long memberId,
+                                                                 MemberProfileImgUpdateRequest memberProfileImgUpdateRequest) {
         String newImageUrl = imageService.uploadSingleImageToS3(memberProfileImgUpdateRequest.getNewProfileImage())
                 .toString();
         memberService.updateMemberProfileImg(memberId, newImageUrl);
-        return new MemberProfileImgUrlResponse(newImageUrl);
+        return new MemberProfileImgUpdateResponse(newImageUrl);
     }
 }
